requirements.txt
aniso8601==9.0.1
apispec==5.1.1
apispec-webframeworks==0.5.2
asttokens==2.0.5
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
black==21.12b0
click==8.0.3
colorama==0.4.4
dateutils==0.6.12
decorator==5.1.1
executing==0.8.2
Flask==2.0.2
flask-restplus==0.13.0
flask-restx==0.5.1
flask-swagger-ui==3.36.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
ipdb==0.13.9
ipython==8.0.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==4.4.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mongoengine==0.23.1
mongoengine-goodjson==1.1.8
mypy-extensions==0.4.3
numpy==1.22.1
packaging==21.3
pandas==1.4.0
parso==0.8.3
pathspec==0.9.0
pep517==0.12.0
pickleshare==0.7.5
pip-tools==6.4.0
platformdirs==2.4.1
pluggy==1.0.0
prompt-toolkit==3.0.24
pure-eval==0.2.2
py==1.11.0
pycodestyle==2.8.0
Pygments==2.11.2
pymongo==3.12.3
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
PyYAML==6.0
six==1.16.0
stack-data==0.1.4
toml==0.10.2
tomli==1.2.3
tqdm==4.62.3
traitlets==5.1.1
typing_extensions==4.0.1
wcwidth==0.2.5
Werkzeug==2.0.2
